+++ src/utils/getClosestPoint.ts
